Author: Karthick and Sam Code refactoring
